Workaround for matmul kernel crash with i8xf32 operands.
The BlockedToMMA pass creates a layout with kWidth=4 when one operand is
i8. However, the TritonGPU to LLVM lowering pass does not support
lowering f32 with kWidth=4, which is the other operand, causing a
segmentation fault.

To work around this, if the operands' minBitWidth is 8 and maxBitWidth
is 32, we use a minBitWidth of 16 instead of 8, creating a layout with
kWidth=2 for both i8 and f32 operands.